move guidance decorator typing to .pyi
